backend/requirements.txt
# Twitter Bot v2.0 Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Browser Automation
playwright>=1.41.0

# AI Integration
anthropic>=0.18.0
google-generativeai>=0.3.0

# Database
aiosqlite>=0.19.0

# Configuration & Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
loguru>=0.7.2

# HTTP Client
httpx>=0.26.0

# WebSocket
websockets>=12.0

# File Import Support
pandas>=2.0.0
openpyxl>=3.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0
